+++ stepper.go
@@ -1,6 +1,9 @@
 package nexus
 
-import "fmt"
+import (
+	"fmt"
+	"strings"
+)
 
 func NewStepper(err *error) *Stepper {
 	return &Stepper{err}
@@ -10,15 +13,19 @@ type Stepper struct {
 	err *error
 }
 
-// Stepf checks internal error before calling the next step. If error is
-// set after step it's wrapped using the given format string, which
-// must contain only one %w.
+// Stepf checks internal error before calling the next step. If error
+// is set after step it's wrapped using the given format string, which
+// must contain one optional %w. If format does not contain %w the
+// format will be extended with ': %w' at the end.
 func (me *Stepper) Stepf(format string, step func()) {
 	if *me.err != nil {
 		return
 	}
 	step()
 	if *me.err != nil {
+		if !strings.Contains(format, "%w") {
+			format += ": %w"
+		}
 		*me.err = fmt.Errorf(format, *me.err)
 	}
 }