+++ changelog.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to semantic versioning.
 
+## [unreleased]
+
+- Stepf takes format with optional %w for wrapping errors
+
 ## [0.6.0] - 2022-08-13
 
 - Add type Writer